Commented issue where error handling should be added.
If parts are added as imaginary builds and they do not have an active
BOM, the script breaks when trying to retrieve the list of missing BOMs.
I'm fine with it breaking for now because it usually just requires
correcting the part number.  Should implement error handling later to
deal with these things.  Especially once this is released for general
use.